Separated data_compilers.js into several files
