+++ frontend/src/js/data_compilers/calendarCompiler.js
@@ -1,0 +1,21 @@
+// Module containing functions which compile the calendar data fetched.
+
+var compileCalendar = function (calendar){
+    //Not implemented yet.
+    return  {
+            events: [
+                {date: "Torsdag den 15 mars", name: "Torsdagspub"},
+                {date: "Torsdag den 15 mars", name: "Fysikalen"},
+                {date: "Torsdag den 15 mars", name: "Torsdagspub"},
+                {date: "Torsdag den 15 mars", name: "Fysikalen"},
+                {date: "Torsdag den 15 mars", name: "Ett väldigt långt namn på event"},
+                {date: "Torsdag den 15 mars", name: "Torsdagspub"},
+                {date: "Torsdag den 15 mars", name: "Fysikalen"},
+                {date: "Torsdag den 15 mars", name: "Torsdagspub"}
+            ]
+        }
+};
+
+module.exports = {
+    compileCalendar
+}
